% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvpaApp.R
\name{mvpaApp}
\alias{mvpaApp}
\title{Starts mvpa.shiny}
\usage{
mvpaApp(run_in_browser = TRUE)
}
\arguments{
\item{run_in_browser}{Shall mvpa.shiny be opened in the default web browser (default = TRUE).}
}
\description{
Starts mvpa.shiny
}
